feat: add legend to columnLine combo
