+++ src/combo/base.ts
@@ -8,6 +8,8 @@ import { isTextUsable } from '../util/common';
 import ThemeController from '../base/controller/theme';
 import { ComboViewConfig } from './util/interface';
 
+const LEGEND_MARGIN = 10;
+
 export interface IComboViewLayer extends ComboViewConfig, LayerConfig {}
 
 export default abstract class ComboViewLayer<T extends IComboViewLayer = IComboViewLayer> extends Layer<T> {
@@ -24,6 +26,7 @@ export default abstract class ComboViewLayer<T extends IComboViewLayer = IComboV
   protected themeController: ThemeController;
   protected geomLayers: ViewLayer[] = [];
   protected colors: string[];
+  protected legends: any[] = [];
 
   constructor(props: T) {
     super(props);
@@ -76,6 +79,10 @@ export default abstract class ComboViewLayer<T extends IComboViewLayer = IComboV
       layer.doDestroy();
     });
     this.geomLayers = [];
+    each(this.legends, (legend) => {
+      legend.destroy();
+    });
+    this.legends = [];
   }
 
   protected createLayer(LayerCtr, data, config) {
@@ -141,9 +148,9 @@ export default abstract class ComboViewLayer<T extends IComboViewLayer = IComboV
     const tickCount = linearScale.ticks.length;
     // 生成右轴ticks
     const max = yAxis.max ? linearScale.max : this.getScaleData(1).max;
-    const tickInterval = max / tickCount;
+    const tickInterval = max / (tickCount - 1);
     const ticks = [];
-    for (let i = 0; i < tickCount + 1; i++) {
+    for (let i = 0; i < tickCount; i++) {
       let tickValue = i * tickInterval;
       if (!Number.isInteger(tickValue)) {
         tickValue = parseFloat(tickValue.toFixed(1));
@@ -163,6 +170,126 @@ export default abstract class ComboViewLayer<T extends IComboViewLayer = IComboV
     const min = values[0];
     const max = yAxis.max ? yAxis.max : values[values.length - 1];
     return { min, max, values };
+  }
+
+  protected getDataByXField(value, index) {
+    const { data, xField } = this.options;
+    const dataSource = data[index];
+    return dataSource.filter((d) => {
+      return d[xField] === value;
+    });
+  }
+
+  protected getYAxisGlobalConfig() {
+    const { min, max, tickCount } = this.options.yAxis;
+    return { min, max, tickCount };
+  }
+
+  protected adjustLayout() {
+    const viewRange = this.getViewRange();
+    const leftPadding = this.geomLayers[0].options.padding;
+    const rightPadding = this.geomLayers[1].options.padding;
+    // 获取legendHeight并加入上部padding
+    let legendHeight = 0;
+    let legendA_BBox;
+    let legendB_BBox;
+    if (this.options.legend?.visible) {
+      legendA_BBox = this.legends[0].get('group').getBBox();
+      legendB_BBox = this.legends[1].get('group').getBBox();
+      legendHeight = legendA_BBox.height + LEGEND_MARGIN * 2;
+    }
+
+    // 同步左右padding
+    const uniquePadding = [leftPadding[0] + legendHeight, rightPadding[1], rightPadding[2], leftPadding[3]];
+    this.geomLayers[0].updateConfig({
+      padding: uniquePadding,
+    });
+    this.geomLayers[0].render();
+    this.geomLayers[1].updateConfig({
+      padding: uniquePadding,
+    });
+    this.geomLayers[1].render();
+    // 更新legend的位置
+    if (this.options.legend?.visible) {
+      this.legends[0].setLocation({
+        x: leftPadding[3] - legendA_BBox.width / 2,
+        y: viewRange.minY + LEGEND_MARGIN,
+      });
+      this.legends[1].setLocation({
+        x: viewRange.maxX - rightPadding[1] - legendB_BBox.width / 2,
+        y: viewRange.minY + LEGEND_MARGIN,
+      });
+    }
+  }
+
+  protected legendFilter(index) {
+    const legend = this.legends[index];
+    const legend_group = legend.get('group');
+    legend_group.on('click', () => {
+      const item = legend.get('items')[0];
+      if (!item.unchecked) {
+        legend.setItemState(item, 'unchecked', true);
+        this.hideLayer(index);
+      } else {
+        legend.setItemState(item, 'unchecked', false);
+        this.showLayer(index);
+      }
+    });
+  }
+
+  protected hideLayer(index) {
+    const layer = this.geomLayers[index];
+    const field = this.options.yField[index];
+    // 隐藏layer时只隐藏yAxis和geometry
+    const { view } = layer;
+    const axisContainer = this.getYAxisContainer(view, field);
+    if (axisContainer) {
+      axisContainer.set('visible', false);
+    }
+    this.setGeometryVisibility(view, false);
+    this.canvas.draw();
+  }
+
+  protected showLayer(index) {
+    const layer = this.geomLayers[index];
+    const field = this.options.yField[index];
+    const { view } = layer;
+    const axisContainer = this.getYAxisContainer(view, field);
+    if (axisContainer) {
+      axisContainer.set('visible', true);
+    }
+    this.setGeometryVisibility(view, true);
+    this.canvas.draw();
+  }
+
+  protected setGeometryVisibility(view, show) {
+    each(view.geometries, (geom) => {
+      const { container, labelsContainer } = geom;
+      if (container) {
+        container.set('visible', show);
+      }
+      if (labelsContainer) {
+        labelsContainer.set('visible', show);
+      }
+    });
+  }
+
+  protected getYAxisContainer(view, field) {
+    let container;
+    const axisCtr = view.controllers.filter((ctr) => {
+      return hasKey(ctr, 'axisContainer');
+    })[0];
+    if (axisCtr) {
+      const ctr = axisCtr as any;
+      const axisGroups = ctr.axisContainer.get('children');
+      each(axisGroups, (g) => {
+        const axisField = g.get('component').get('field');
+        if (axisField === field) {
+          container = g;
+        }
+      });
+    }
+    return container;
   }
 
   protected drawTitle(): void {
